requirements.txt
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
fiona==1.10.1
geopandas==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
shapely==2.1.1
six==1.17.0
tzdata==2025.2
python-dotenv
